update automated port scanning

+++ src/toolkit/portscan/port_loader.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PortScan {
     pub protocol: String,
-    pub port: u16,
+    pub port: u32,
     pub description: String,
 }
